+++ core/model/src/main/java/com/forknowledge/feature/model/Nutrient.kt
@@ -4,5 +4,5 @@ data class Nutrient(
     val name: String = "",
     val amount: Double = 0.0,
     val unit: String = "",
-    val percentOfDailyNeeds: Long = 0
+    val percentOfDailyNeeds: Double = 0.0
 )
